Fix imports & exports in workflows
