+++ src/ConvergenceTestWorkflow/ConvergenceTestWorkflow.jl
@@ -1,13 +1,5 @@
 module ConvergenceTestWorkflow
 
-using ExpressBase: Scf, calculation
-
-function isconvergent(a::AbstractVector)
-    terms = abs.(diff(a))
-    x, y, z = last(terms, 3)
-    return all(0 <= r < 1 for r in (y / x, z / y))
-end
-
 include("Config.jl")
 include("actions.jl")
 include("Recipes.jl")
