+++ src/EquationOfStateWorkflow/EquationOfStateWorkflow.jl
@@ -1,13 +1,5 @@
 module EquationOfStateWorkflow
 
-using ExpressWorkflowMaker.Templates: LogTime, RunCmd
-
-using ..Express: current_software
-
-export MakeInput, FitEos, GetData, SaveEos, RunCmd, LogTime, calculation
-
-const ScfOrOptim = Union{SelfConsistentField,Optimization}
-
 include("Config.jl")
 include("actions.jl")
 include("Recipes.jl")
