+++ src/PhononWorkflow/PhononWorkflow.jl
@@ -1,23 +1,5 @@
 module PhononWorkflow
 
-using AbInitioSoftwareBase: load
-using AbInitioSoftwareBase.Inputs: Input
-using ExpressWorkflowMaker.Templates: LogTime, RunCmd
-using SimpleWorkflows.Workflows: Workflow, run!, →
-using Unitful: ustrip, @u_str
-
-using ExpressBase:
-    Calculation,
-    LatticeDynamics,
-    SelfConsistentField,
-    Scf,
-    FixedIonSelfConsistentField,
-    VariableCellOptimization,
-    Action
-using ..Express: current_software
-
-export Dfpt, MakeInput, LogTime, run!
-
 include("Config.jl")
 include("actions.jl")
 include("Recipes.jl")
